Fixed issue with websocket connection closing when the tab is in the background.
With the tab in a background, timers may be delayed, leading to the timer that collects data to sent to fire late. This resulted in a too large websocket message being sent with Chrome because there was more audio data accumulated. Fix it by limiting the size of messages and generating as many messages as needed.